requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.5
Django==4.2.6
djangorestframework==3.14.0
jwt==1.3.1
mysqlclient==2.2.0
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
